Driving
Got the rovers driving on start and accepting commands.  Communication
is still one way and must be done from the rover command line in the
interface with all necassary tags.  Also converted RoverAutonomusCode to
abstract and created a template for extending it into new classes to be
loaded with autonomous processing.